--- app/interactions/client_hold_interaction/base.rb
@@ -1,8 +1,0 @@
-module ClientHoldInteraction
-  class Base < ActiveInteraction::Base
-    def to_stdout(message)
-      time = Time.zone.now.utc
-      STDOUT << "#{time} - #{message}\n" unless Rails.env.test?
-    end
-  end
-end
